# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TreeKB API
  version: "2.3.0"
  description: |
    TreeKB is an AI-driven, self-evolving knowledge graph system. It ingests documents,
    extracts knowledge points (KPs), organises them into a hierarchical topic tree, and
    provides intelligent question-answering with citations.

    **Key concepts**
    - **Dataset** -- an isolated knowledge base with its own SQLite database.
    - **Node** -- a topic node in the hierarchical tree.
    - **Chunk / KP** -- an atomic knowledge point stored under a node.
    - **Schema node** -- a curated node that guides ingestion mapping.
    - **Decision** -- a pending human-review item produced by the KP decision engine.

    Most endpoints require the `X-Dataset-ID` header to select the active dataset.
    Endpoints under `/datasets`, `/health`, `/metrics`, and `/settings` do not.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  parameters:
    DatasetIdHeader:
      name: X-Dataset-ID
      in: header
      required: true
      description: UUID of the target dataset. Resolved by middleware; falls back to default dataset when omitted.
      schema:
        type: string
        format: uuid
  schemas:
...
    SimpleError:
      type: object
      properties:
        error:
          type: string

    # ── Dataset ───────────────────────────────────────────────────────────────
...
    Node:
      type: object
      properties:
        node_id:
          type: string
        name:
          type: string
        parent_id:
          type: string
          nullable: true
        level:
          type: integer
        node_summary:
          type: string
          nullable: true
        node_description:
          type: string
          nullable: true
        is_schema_node:
          type: integer
          enum: [0, 1]
        aliases_json:
          type: string
        keywords_json:
          type: string
        scope_json:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Chunk ─────────────────────────────────────────────────────────────────
    Chunk:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        node_id:
          type: string
        document_id:
          type: integer
          nullable: true
        doc_title:
          type: string
        kp_type:
          type: string
          enum: [fact, rule, definition, procedure, example, context, paragraph_context, legacy_chunk]
        chunk_type:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Document ──────────────────────────────────────────────────────────────
...
paths:
...
  /nodes:
    get:
      operationId: getFullTree
      tags: [Nodes]
      summary: Get full tree structure
      description: Returns the complete hierarchical tree of all nodes along with tree statistics.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Full tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                    type: array
                    items:
                      type: object
                      description: Nested node objects with children arrays
                  stats:
                    type: object
                    properties:
                      total_nodes:
                        type: integer
                      total_chunks:
                        type: integer
                      max_depth:
                        type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    post:
      operationId: createNode
      tags: [Nodes]
      summary: Create a new node
      description: Creates a new node in the knowledge tree.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [node_id, name]
              properties:
                node_id:
                  type: string
                  description: Unique node identifier
                name:
                  type: string
                parent_id:
                  type: string
                  nullable: true
                summary:
                  type: string
                scope:
                  type: object
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
...
  /nodes/search:
    get:
      operationId: searchNodes
      tags: [Nodes]
      summary: Search chunks by content
      description: Performs a BM25 full-text search across all chunks and groups results by their parent node.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: q
          in: query
          required: true
          description: Search query text
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum chunks to return (1-100)
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Search results grouped by node
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        node_id:
                          type: string
                        node_name:
                          type: string
                        chunks:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                              preview:
                                type: string
                              score:
                                type: number
                  total_chunks:
                    type: integer
        "400":
          description: Missing q parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [Nodes]
      summary: Get node by ID
      description: Returns a single node. Pass `?context=true` to include parent chain and sibling context.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: context
          in: query
          description: Include parent chain and sibling context
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    put:
      operationId: updateNode
      tags: [Nodes]
      summary: Update a node
      description: Updates one or more fields of a node (name, summary, scope, aliases, description, parent_id for reparenting).
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                summary:
                  type: string
                scope:
                  type: object
                aliases:
                  type: array
                  items:
                    type: string
                node_description:
                  type: string
                parent_id:
                  type: string
                  nullable: true
                  description: Set to a node ID to reparent, or null to make root-level
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          description: No updates provided or invalid parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "404":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    delete:
      operationId: deleteNode
      tags: [Nodes]
      summary: Delete a node
      description: |
        Deletes a node and all its chunks. Children are reparented to the deleted node's parent.
        The root node cannot be deleted.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  name:
                    type: string
                  chunksDeleted:
                    type: integer
                  childrenReparented:
                    type: integer
        "400":
          description: Cannot delete root node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "404":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /nodes/{id}/children:
    get:
      operationId: getNodeChildren
      tags: [Nodes]
      summary: Get node children
      description: Returns the direct children of a node.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Children list
          content:
            application/json:
              schema:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/Node"
                  count:
                    type: integer
  /nodes/{id}/entities:
    get:
      operationId: getNodeEntities
      tags: [Nodes]
      summary: Get entities and facts for a node
      description: Returns entities and their facts that are associated with chunks under a specific node.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum entities (1-500)
          schema:
            type: integer
            default: 50
        - name: debug
          in: query
          description: Include debug information
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
      responses:
        "200":
          description: Entities and facts
          content:
            application/json:
              schema:
                type: object
  /nodes/{id}/aliases:
    post:
      operationId: generateNodeAliases
      tags: [Nodes]
      summary: Generate aliases for a node
      description: Uses LLM to generate search aliases for a specific node (e.g. abbreviations, synonyms).
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated aliases
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  nodeId:
                    type: string
                  aliases:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /nodes/{id}/chunks:
    post:
      operationId: createChunk
      tags: [Nodes]
      summary: Create a manual chunk on a node
      description: Adds a manually-created knowledge point to a specific node.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: The knowledge point content
                kp_type:
                  type: string
                  enum: [fact, rule, definition, procedure, example, context]
                  default: fact
                doc_title:
                  type: string
                  default: Manual Entry
      responses:
        "200":
          description: Chunk created
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunk:
                    type: object
                    properties:
                      id:
                        type: integer
                      content:
                        type: string
                      kp_type:
                        type: string
                      doc_title:
                        type: string
        "400":
          description: Missing content or invalid kp_type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "404":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /aliases/sync:
    post:
      operationId: syncAliases
      tags: [Nodes]
      summary: Generate aliases for all nodes missing them
      description: Batch generates LLM-based aliases for nodes that do not yet have any.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  default: 50
                  description: Maximum nodes to process
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: integer
                  total_aliases:
                    type: integer
  /chunks/{nodeId}:
    get:
      operationId: getChunksForNode
      tags: [Nodes]
      summary: Get chunks for a node
      description: Returns all active chunks assigned to a specific node.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chunk list
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Chunk"
                  count:
                    type: integer
  /chunks/detail/{id}:
    get:
      operationId: getChunkById
      tags: [Nodes]
      summary: Get chunk by ID
      description: Returns a single chunk with all fields.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chunk details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chunk"
        "404":
          description: Chunk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /chunks/{id}:
    delete:
      operationId: deleteChunk
      tags: [Nodes]
      summary: Delete a manual chunk
      description: Deletes a manually-created chunk. File-ingested chunks cannot be deleted via this endpoint (returns 403).
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chunk deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "403":
          description: Cannot delete file-ingested chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "404":
          description: Chunk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

  # ── Entities ──────────────────────────────────────────────────────────────
...